refactor(frr): extract prefix-list rendering to sibling file
Move the FRR `ip|ipv6 prefix-list` rendering helpers out of the
2,470-LOC policy_render.go into a cohesive sibling prefix_list_render.go
(#6424). This is the first of several pure code-motion extractions that
break the god-renderer into per-aspect files.

Moved whole, unchanged: renderRouteFilterEntry, renderFromPrefixListACL,
the indexedRouteFilter type + partitionRouteFiltersByFamily, and the
fromPrefixListRef type + fromPrefixListRefs. No signatures, logic, or
render output change; the emitted frr.conf is byte-identical.

Validation: go build ./pkg/frr, gofmt, go vet clean. policy_render.go
drops from 2,470 to 2,202 LOC.
